+++ CHANGELOG.md
@@ -76,6 +76,7 @@ doivent être mises à jour #1667
 - Suppression de HeliosGeneriqueXMLFile, TdTExtractionHelios #2211
 - Suppression de IParapheur::getMetaDonnee et SignatureRecuperation::getMetaDonnee #2211
 - Suppression de FluxDataStandard, use FluxDataSedaDefault instead #2211
+- Suppression de SignatureConnecteur::getDossierID($id, $name) #2289
 
 ## Dépréciations